Fixed errors in build. Unified styling of project files
