+++ libraries/Microsoft.Bot.Builder.Prompts/Choices/Choice.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Collections.Generic;
 using Microsoft.Bot.Schema;
+using System.Collections.Generic;
 
 namespace Microsoft.Bot.Builder.Prompts.Choices
 {
