+++ libraries/Microsoft.Bot.Builder.Prompts/Choices/ModelResult.cs
