+++ libraries/Microsoft.Bot.Builder.Prompts/Choices/RecognizeChoices.cs
@@ -1,10 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Recognizers.Text.Number;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Recognizers.Text.Number;
 
 namespace Microsoft.Bot.Builder.Prompts.Choices
 {
@@ -18,8 +18,8 @@ namespace Microsoft.Bot.Builder.Prompts.Choices
         public static List<ModelResult<FoundChoice>> RecognizeChoices(string utterance, List<Choice> list, FindChoicesOptions options = null)
         {
             // Try finding choices by text search first
-            // - We only want to use a single strategy for returning results to avoid issues where utterances 
-            //   like the "the third one" or "the red one" or "the first division book" would miss-recognize as 
+            // - We only want to use a single strategy for returning results to avoid issues where utterances
+            //   like the "the third one" or "the red one" or "the first division book" would miss-recognize as
             //   a numerical index or ordinal as well.
 
             var locale = options?.Locale ?? Recognizers.Text.Culture.English;
@@ -96,6 +96,7 @@ namespace Microsoft.Bot.Builder.Prompts.Choices
                     }
                 }).ToList();
         }
+
         private static List<ModelResult<FoundChoice>> RecognizeNumber(string utterance, string culture)
         {
             var model = new NumberRecognizer(culture).GetNumberModel(culture);
