+++ libraries/Microsoft.Bot.Builder.Prompts/Choices/SortedValue.cs
