+++ libraries/Microsoft.Bot.Builder.Prompts/NumberPrompt.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.Bot.Schema;
 using Microsoft.Recognizers.Text;
 using Microsoft.Recognizers.Text.Number;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 using static Microsoft.Bot.Builder.Prompts.PromptValidatorEx;
 
 namespace Microsoft.Bot.Builder.Prompts
@@ -36,7 +36,7 @@ namespace Microsoft.Bot.Builder.Prompts
     }
 
     /// <summary>
-    /// NumberPrompt recognizes floats or ints 
+    /// NumberPrompt recognizes floats or ints
     /// </summary>
     public class NumberPrompt<T> : BasePrompt<NumberResult<T>>
     {
@@ -56,8 +56,8 @@ namespace Microsoft.Bot.Builder.Prompts
 
         /// <summary>
         /// Used to validate the incoming text, expected on context.Activity, is
-        /// valid according to the rules defined in the validation steps. 
-        /// </summary>        
+        /// valid according to the rules defined in the validation steps.
+        /// </summary>
         public override async Task<NumberResult<T>> Recognize(ITurnContext context)
         {
             BotAssert.ContextNotNull(context);
