+++ libraries/Microsoft.Bot.Builder.Prompts/TextPrompt.cs
@@ -1,6 +1,7 @@
-﻿using System;
+﻿using Microsoft.Bot.Schema;
+using System;
 using System.Threading.Tasks;
-using Microsoft.Bot.Schema;
+
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
@@ -35,12 +36,12 @@ namespace Microsoft.Bot.Builder.Prompts
     /// <summary>
     /// Represents a user prompt class for text input.
     /// </summary>
-    /// <remarks>The <see cref="Recognize(ITurnContext)"/> method passes any 
+    /// <remarks>The <see cref="Recognize(ITurnContext)"/> method passes any
     /// non-whitespace string to the custom validator, if one was provided.
-    /// To change this behavior, derive from this class and add your own custom 
-    /// validation behavior. 
-    /// <para>For simple validation changes, specify a <see cref="PromptValidator{InT}"/> 
-    /// in the constructor. If the standard validation passes, the custom 
+    /// To change this behavior, derive from this class and add your own custom
+    /// validation behavior.
+    /// <para>For simple validation changes, specify a <see cref="PromptValidator{InT}"/>
+    /// in the constructor. If the standard validation passes, the custom
     /// validator is called on the recognized value.</para>
     /// </remarks>
     public class TextPrompt : BasePrompt<TextResult>
@@ -50,7 +51,7 @@ namespace Microsoft.Bot.Builder.Prompts
         /// </summary>
         /// <param name="validator">The input validator for the prompt object.</param>
         /// <remarks><paramref name="validator"/> is called only if the
-        /// <see cref="Recognize(ITurnContext)"/> method recognizes a value. 
+        /// <see cref="Recognize(ITurnContext)"/> method recognizes a value.
         /// </remarks>
         public TextPrompt(PromptValidator<TextResult> validator = null)
             : base(validator)
@@ -64,7 +65,7 @@ namespace Microsoft.Bot.Builder.Prompts
         /// <returns>A task that represents the work queued to execute.</returns>
         /// <remarks>Call this when you expect that the incoming activity for this
         /// turn contains the user input to recognize.
-        /// If recognition succeeds, the <see cref="TextResult.Value"/> property of the 
+        /// If recognition succeeds, the <see cref="TextResult.Value"/> property of the
         /// result contains the value recognized.
         /// <para>If recognition fails, returns a <see cref="TextResult"/> with
         /// its <see cref="PromptStatus"/> set to <see cref="PromptStatus.NotRecognized"/> and
@@ -88,6 +89,5 @@ namespace Microsoft.Bot.Builder.Prompts
             }
             return textResult;
         }
-
     }
 }
